api/requirements.txt
Flask
Flask-Swagger-Ui
Flask-Migrate
Flask-Cors
python-dotenv